requirements.txt
altair==3.3.0
dash==2.0.0
dash_dangerously_set_inner_html==0.0.2
dash_daq==0.5.0
dash-bootstrap-components==1.0.0
dash-extensions==0.0.63
dash-leaflet >= 0.1.18
email_validator==1.1.3
Flask==2.0.1
flask-bcrypt==0.7.1
flask-login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
folium==0.12.1
geobuf==1.1.1
geojson==2.5.0
geopandas==0.9.0
jsbeautifier==1.14.0
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
plotly==5.3.1
pyarrow==5.0.0
pygeos==0.10.2
rtree==0.9.7
scipy==1.7.1
shapely == 1.8a3
utm==0.7.0
WTForms==2.3.3
xlrd==2.0.1
psycopg2-binary
GeoAlchemy2==0.10.2
Werkzeug==2.0.0
protobuf==3.20.0